requirements.txt
# Core dependencies
playwright==1.40.0
python-telegram-bot==20.7
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.0
aiosqlite==0.19.0

# AI and text processing
openai==1.6.1
nltk==3.8.1
textblob==0.17.1

# Utilities
aiohttp==3.9.1
asyncio==3.4.3
pydantic==2.5.2
loguru==0.7.2
fake-useragent==1.4.0
pytz==2023.3
schedule==1.2.0

# Data processing
pandas==2.1.4
numpy==1.26.2
openpyxl==3.1.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0